# swagger/swagger.yml
swagger: "2.0"
info:
  title: beego Test API
  description: beego has a very cool tools to autogenerate documents for your API
  version: 1.0.0
  termsOfService: http://beego.me/
  contact:
    email: astaxie@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
basePath: /v1
paths:
  /account/:
    get:
      tags:
      - account
      description: get AdminAccounts
      operationId: AdminAccountsController.Get List Acount
      parameters:
      - in: header
        name: Authorization
        description: token
        required: true
        type: string
      - in: query
        name: hotelId
        description: hotel id
        type: integer
        format: int64
      - in: query
        name: fullName
        description: account full name
        type: string
      - in: query
        name: email
        description: account email
        type: string
      - in: query
        name: status
        description: account status
        type: string
      - in: query
        name: fields
        description: Default *.Fields returned. e.g. col1,col2 ...
        type: string
      - in: query
        name: limit
        description: Limit the size of result set. Must be an integer
        type: string
      - in: query
        name: page
        description: Start position of result set. Must be an integer
        type: string
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/models.SwaggerListAccount'
        "403":
          description: '101 : Permission denied <br> 104 : Invalid Token <br>'
    post:
      tags:
      - account
      description: create AdminAccounts
      operationId: AdminAccountsController.Create Admin Account
      parameters:
      - in: header
        name: Authorization
        description: Bearer token
        required: true
        type: string
      - in: body
        name: body
        description: body for AdminAccounts content
        required: true
        schema:
          $ref: '#/definitions/models.InsertAccountFields'
      responses:
        "200":
          description: Add new account success
          schema:
            $ref: '#/definitions/models.SwaggerDetailAccount'
        "403":
          description: '101 : Permission denied <br> 104 : Invalid Token <br> 224
            : Fields value not exists <br> 221 : Parse json body false <br> 215 :
            Incorrect password format (Password must contain at least eight characters
            to 60 characters and uppercase and number) <br> 217 : Field required <br>
            213 : Invalid Email format <br> 210 : Field not numeric <br> 301 : Account
            Exists <br> 303 : Save Database false'
  /account/{id}:
    get:
      tags:
      - account
      description: get AdminAccounts by id
      operationId: AdminAccountsController.Detail
      parameters:
      - in: header
        name: Authorization
        description: Bearer token
        required: true
        type: string
      - in: path
        name: id
        description: The key for staticblock
        required: true
        type: string
      responses:
        "200":
          description: Get Success
          schema:
            $ref: '#/definitions/models.SwaggerDetailAccount'
        "403":
          description: '101 : Permission denied <br> 104 : Invalid Token <br> 302
            : Account not found'
    put:
      tags:
      - account
      description: update the AdminAccounts
      operationId: AdminAccountsController.Update Account
      parameters:
      - in: header
        name: Authorization
        description: Bearer token
        required: true
        type: string
      - in: path
        name: id
        description: The id you want to update
        required: true
        type: string
      - in: body
        name: body
        description: body for AdminAccounts content
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccountFields'
      responses:
        "200":
          description: Update account success
          schema:
            $ref: '#/definitions/models.AddAdminAccountStruct'
        "403":
          description: '101 : Permission denied <br> 104 : Invalid Token <br> 224
            : Fields value not exists <br> 221 : Parse json body false <br> 215 :
            Incorrect password format (Password must contain at least eight characters
            to 60 characters and uppercase and number) <br> 217 : Field required <br>
            213 : Valid Email format <br> 210 : Field not numeric <br> 301 : Email
            Exists <br> 303 : Save Database false'
    delete:
      tags:
      - account
      description: delete the AdminAccounts
      operationId: AdminAccountsController.Delete
      parameters:
      - in: header
        name: Authorization
        description: Bearer token
        required: true
        type: string
      - in: path
        name: id
        description: The id you want to delete
        required: true
        type: string
      responses:
        "200":
          description: delete success!
          schema:
            $ref: '#/definitions/libs.ResponseJson'
        "403":
          description: '101 : Permission denied <br> 104 : Invalid Token <br> 302
            : Account Not Found <br> 303 : Delete False'
...
definitions:
...
  2798.0xc0006265d0.false:
    title: "false"
    type: object
  2852.0xc000626600.false:
    title: "false"
    type: object
...
  libs.ResponseJson:
    title: ResponseJson
    type: object
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/2798.0xc0006265d0.false'
      detailErrorCode:
        $ref: '#/definitions/2852.0xc000626600.false'
      message:
        type: string
      success:
        type: string
  models.AddAdminAccountStruct:
    title: AddAdminAccountStruct
    type: object
    properties:
      Email:
        type: string
      FullName:
        type: string
      IsRoot:
        type: integer
        format: int32
      RoleId:
        type: integer
        format: int64
      Status:
        type: integer
        format: int32
...
  models.AdminAccounts:
    title: AdminAccounts
    type: object
    properties:
      CreatedAt:
        type: string
        format: datetime
      CreatedUser:
        type: integer
        format: int64
      DeletedAt:
        type: integer
        format: int32
      Email:
        type: string
      FullName:
        type: string
      Id:
        type: integer
        format: int64
      IsRoot:
        type: integer
        format: int32
      Password:
        type: string
      RoleId:
        type: integer
        format: int64
      Status:
        type: integer
        format: int32
      UpdatedAt:
        type: string
        format: datetime
      UpdatedUser:
        type: integer
        format: int64
...
  models.InsertAccountFields:
    title: InsertAccountFields
    type: object
    properties:
      Email:
        type: string
      FullName:
        type: string
      Password:
        type: string
      RoleId:
        type: integer
        format: int64
      Status:
        type: integer
        format: int32
...
  models.SwaggerDetailAccount:
    title: SwaggerDetailAccount
    type: object
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/models.AdminAccounts'
      detailErrorCode:
        type: array
        items:
          type: string
      message:
        type: string
      success:
        type: string
...
  models.SwaggerListAccount:
    title: SwaggerListAccount
    type: object
    properties:
      code:
        type: string
      data:
        type: array
        items:
          $ref: '#/definitions/models.AdminAccounts'
      detailErrorCode:
        type: array
        items:
          type: string
      message:
        type: string
      success:
        type: string
...
  models.UpdateAccountFields:
    title: UpdateAccountFields
    type: object
    properties:
      Email:
        type: string
      FullName:
        type: string
      Password:
        type: string
      RoleId:
        type: integer
        format: int64
      Status:
        type: integer
        format: int32
...